backend/requirements.txt
django~=4.1.3
pillow
django-debug-toolbar
djangorestframework
django-cors-headers